requirements.txt
pygame-ce>=2.4.0
openai>=1.30.5
anthropic>=0.28.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0
PyOpenGL>=3.1.7
numpy>=1.26.4
pyyaml>=6.0.1
requests>=2.32.3
python-dotenv>=1.0.1
pyperclip>=1.8.0
pywinpty>=2.0.0